Update time 2020-2-3 21:30,Update V1.0.3.2

+++ MostBrain/src/cn/epicfx/winfxk/mostbrain/tool/Tool.java
@@ -291,7 +291,7 @@ public class Tool implements X509TrustManager, HostnameVerifier {
 		if (obj == null || string.isEmpty())
 			return d;
 		try {
-			return Long.valueOf(string);
+			return Long.valueOf(Float.valueOf(string).intValue());
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
